requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
emoji==0.6.0
idna==2.10
Pillow==8.0.1
pip-autoremove==0.9.1
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.4
python-i18n==0.3.9
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
